feat: adding discord logo to auth flow
