+++ account-kit/react/src/components/auth/sections/OAuth.tsx
@@ -1,5 +1,10 @@
 import { memo } from "react";
-import { AppleIcon, FacebookIcon, GoogleIcon } from "../../../icons/oauth.js";
+import {
+  AppleIcon,
+  FacebookIcon,
+  GoogleIcon,
+  DiscordIcon,
+} from "../../../icons/auth-icons/index.js";
 import { assertNever } from "../../../utils.js";
 import { Button } from "../../button.js";
 import { useOAuthVerify } from "../hooks/useOAuthVerify.js";
@@ -33,11 +38,7 @@ export const OAuth = memo(({ ...config }: Props) => {
       );
     case "discord":
       return (
-        <Button
-          variant="social"
-          icon={<div>Discord</div>}
-          onClick={authenticate}
-        >
+        <Button variant="social" icon={<DiscordIcon />} onClick={authenticate}>
           Discord
         </Button>
       );
