+++ account-kit/react/src/icons/auth-icons/index.ts
@@ -1,0 +1,6 @@
+import { DiscordIcon } from "./discord.js";
+import { FacebookIcon } from "./facebook.js";
+import { GoogleIcon } from "./google.js";
+import { AppleIcon } from "./apple.js";
+
+export { DiscordIcon, GoogleIcon, FacebookIcon, AppleIcon };
