+++ examples/ui-demo/src/components/icons/discord.tsx
@@ -1,0 +1,45 @@
+import { JSX, SVGProps } from "react";
+
+export const DiscordLogo = ({
+  fill,
+  ...props
+}: JSX.IntrinsicAttributes & SVGProps<SVGSVGElement>) => (
+  <svg
+    width="21"
+    height="20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <g clip-path="url(#a)">
+      <path
+        d="M15.9785 0H5.35352C2.76468 0 .666016 2.09867.666016 4.6875v10.625C.666016 17.9013 2.76468 20 5.35352 20H15.9785c2.5889 0 4.6875-2.0987 4.6875-4.6875V4.6875C20.666 2.09867 18.5674 0 15.9785 0Z"
+        fill="#5865F2"
+      />
+      <mask
+        id="b"
+        maskUnits="userSpaceOnUse"
+        x="2"
+        y="3"
+        width="17"
+        height="14"
+      >
+        <path
+          d="M2.85352 3.98438H18.4785V16.0883H2.85352V3.98438Z"
+          fill="#fff"
+        />
+      </mask>
+      <g mask="url(#b)">
+        <path
+          d="M16.0807 5.06249c-1.013-.46502-2.0821-.79662-3.1804-.98649-.01-.00185-.0203-.0005-.0295.00385s-.0167.01147-.0216.02037c-.1374.24422-.2896.56297-.396.81344-1.2011-.17985-2.3958-.17985-3.57215 0-.10657-.2561-.26422-.56922-.40219-.81344-.00511-.00871-.01272-.01568-.02185-.02-.00913-.00433-.01935-.00579-.02932-.00422-1.09843.18949-2.16754.5211-3.18039.98649-.00867.00368-.01599.00995-.02094.01796C3.20073 8.1067 2.64589 11.0586 2.91808 13.9739c.00075.0071.00294.0141.00643.0203.00349.0063.00821.0118.01388.0163 1.33641.9814 2.63102 1.5773 3.90156 1.9722.00989.003.02043.0028.03024-.0004.0098-.0032.01838-.0093.02461-.0175.30054-.4105.56843-.8433.7982-1.2983.00316-.0063.00497-.0131.0053-.0201.00033-.007-.00082-.0139-.00339-.0204-.00256-.0065-.00647-.0124-.01147-.0173-.00501-.0049-.01099-.0087-.01755-.0111-.425-.1612-.82961-.3577-1.21883-.5809-.00709-.0042-.01304-.01-.01733-.017-.00429-.0071-.00679-.015-.00728-.0232-.00049-.0083.00106-.0165.00449-.0239.00343-.0075.00865-.014.0152-.019.08211-.0614.1628-.1246.24203-.1897.00694-.0057.01534-.0094.02425-.0106.00892-.0013.018 0 .02622.0037 2.55719 1.1675 5.32556 1.1675 7.85246 0 .0083-.0039.0175-.0054.0265-.0043.0091.0011.0176.0048.0247.0106.0793.0654.1602.1289.2427.1903.0065.0049.0118.0114.0153.0188.0035.0075.0051.0157.0047.0239-.0005.0082-.0029.0162-.0071.0232-.0043.0071-.0102.013-.0172.0172-.3895.2274-.7974.4215-1.2195.5803-.0065.0025-.0125.0064-.0174.0113-.005.005-.0089.011-.0114.0175-.0025.0066-.0036.0136-.0032.0206.0004.007.0023.0138.0055.0201.2338.4514.5003.8852.7975 1.2976.0061.0085.0146.0148.0245.0181.0098.0034.0204.0035.0304.0005 1.2767-.3949 2.5713-.9908 3.9078-1.9722.0057-.0043.0105-.0096.014-.0158.0035-.0062.0056-.0131.0063-.0202.3258-3.3704-.5457-6.29811-2.3102-8.89342-.0044-.00843-.0116-.01503-.0203-.01859ZM8.07487 12.1988c-.76992 0-1.40429-.7068-1.40429-1.5748 0-.86808.62211-1.57487 1.40429-1.57487.78828 0 1.41649.71304 1.40414 1.57487 0 .868-.62195 1.5748-1.40414 1.5748Zm5.19183 0c-.7699 0-1.4042-.7068-1.4042-1.5748 0-.86808.6221-1.57487 1.4042-1.57487.7883 0 1.4165.71304 1.4043 1.57487 0 .868-.616 1.5748-1.4043 1.5748Z"
+          fill="#fff"
+        />
+      </g>
+    </g>
+    <defs>
+      <clipPath id="a">
+        <path fill="#fff" transform="translate(.666016)" d="M0 0h20v20H0z" />
+      </clipPath>
+    </defs>
+  </svg>
+);
